+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.10] - 2026-03-18
+
+### Added
+
+* New module `src/postfix_logsums/functions.py`.
+* New module `src/postfix_logsums/parser.py`.
+
+### Changed
+
+* Moving common used functions into `src/postfix_logsums/functions.py`.
+* Moving class `PostfixLogParser` into module `src/postfix_logsums/parser.py`.
+* Simplyfication of tests.
+* Renaming RPM package to postfix-logsums.
+
 ## [0.9.9] - 2026-03-17
 
 ### Added
@@ -94,6 +108,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Initial release
 
+[0.9.10]: https://github.com/pixelpark/postfix-logsums/compare/0.9.9...0.9.10
 [0.9.9]: https://github.com/pixelpark/postfix-logsums/compare/0.9.4...0.9.9
 [0.9.4]: https://github.com/pixelpark/postfix-logsums/compare/0.9.3...0.9.4
 [0.9.3]: https://github.com/pixelpark/postfix-logsums/compare/0.9.2...0.9.3